.gitmodules: always use our own submodule repos on GitHub.
The driving factor for this is that the CentOS builder is so
old that its copy of curl/OpenSSL doesn't support SNI properly,
so trying to clone from xiph.org fails (since the default certificate
for the site specifies opus-codec.org, not xiph.org).

It's also a bit nicer not to leech on Xiph.org's bandwidth. Plus,
we now have the possibility to do fixups (that aren't upstreamable for
some reason) to these repos without too much hassle.