ConnectDialog: Handle HTTP URLs from clipboard
Implements feature request #3143

In various communication programs when a user writes a Mumble server
address without the mumble:// schema, the program will interpret a
domain example.com as a HTTP hyperlink and convert it to a HTTP
hyperlink. Some programs will only make the text as is a hyperlink,
others will also convert the written text to include http (which will
make it more confusing for users who are not familiar with HTTP, URLs
and schemas/the Mumble schema).

Instead of pre-filling the add server dialog, introduce a
paste-able-clipboard notice to the ConnectDialogEdit that suggests to
the user to pre-fill with the clipboard content.