scripts: add generate-ffdhe.py for generating FFDHETable.h.
This commit adds a new script for generating PEM forms
of the Diffie-Hellman groups in RFC 7919.

RFC 7919 specifies a set of Diffie-Hellman groups that
can be negotiated by TLS 1.1 and 1.2 servers and clients
using the same mechanism that is used for named elliptic
curves in TLS.

Our TLS implementation doesn't support RFC 7919 at present,
but we can still use the groups defined in the RFC as good
reference Diffie-Hellman parameters for Murmur, since they
are audited and may be more resistant to attacks than ones
randomly generated.