FFDHE: new class for accessing RFC 7919 FFDHE parameters.
This commit adds a new class that gives access to the
Diffie-Hellman parameters from RFC 7919.

The class exposes a static function, PEMForNamedGroup, which
looks up a set of Diffie-Hellman parameters in PEM form by
the name used in RFC 7919.

For example, to get the PEM form of 'ffdhe2048', you would
call

	QByteArray pem = FFDHE::PEMForNamedGroup(QLatin1String("ffdhe2048"));

With that in hand, you can pass the returned PEM data to
QSslDiffieHellmanParameters to construct an object that can be set
on a QSslSocket-based TLS server.