Overlay_win: properly terminate our overlay helper processes.
This gets rid of the errors in the Mumble log about destroying
QProcess instances while they're still running.

The QProcess instances in OverlayPrivateWin are parented to the
OverlayPrivateWin class.

The warning happened because the QObject destructor for OverlayPrivateWin
would destroy the QProcesses, because they are its children. However, at
that point, the processes were still running.

This commit remedies that by terminating the processes before the QObject
destructor runs.