GlobalShortcutWin, MumbleApplication: inject native WM_* keyboard and mouse messages into GlobalShortcutWin.
This should fix the problem where some keys could not be observed in
Mumble when Mumble was the program in focus. People have reported this
problem with keys F13-F24.

The reason these keys could not be observed in Mumble is that:

1) Our keyboard and mouse hook callbacks do not receive keypresses and
   mouse clicks when Mumble is on focus. We only recieve events when
   other programs are focused.

2) DirectInput does not know of F13-F24.

To remedy this situation, this commit implements a system where we feed
the native Windows keyboard and mouse events from Qt's nativeEventFilter
into GlobalShortcutWin.

Technically, we provide methods in GlobalShortcutWin that allows the
nativeEventFilter to inject keyboard and mouse evnets into it. The events
from nativeEventFilter are received on the main thread. To inject them
into GlobalShortcutWin, we inject the mouse and keyboard messages through
two QEvents: InjectKeyboardMessageEvent and InjectMouseMessageEvent.

GlobalShortcutWin implements two new public-facing methods:
injectKeyMessage and injectMouseMessage. Internally, these methods
will cause a InjectKeyboardMessageEvent or InjectMouseMessageEvent
to be sent to the GlobalShortcutWin's event handler. This is done
through QCoreApplication::postEvent(), which means that the event
is handled in the receiving object's thread.

To achieve the ability to inject these events, some of general keyboard
and mouse handling code from our hook callbacks HookKeyboard and HookMouse
have been refactored into two new methods, handleKeyMessage and
handleMouseMessage.

These new methods are used in both HookKeyboard and the event handler for
InjectKeyboardMessageEvent, as well as in HookMouse and the event handler
for InjectMouseMessageEvent.

Note: the current implementation does not implement suppression of
keyboard and mouse events coming from the nativeEventFilter. This would
require us to block the main thread until GlobalShortcutWin has time to
handle the event. This is considered future work.