GlobalShortcutWin: add comment about button indexes in HookMouse().
Document that the button indexes in HookMouse() are DirectInput button
indexes.

The reason we use those, is that Mumble originally only used DirectInput
for global shortcuts on Windows. Later on, we added support for keyboard
and mouse hooks to get global keypresses and mouse clicks.

When these keyboard and mouse hooks are used, we actually get the same
event from *both* sources: First, we get it from the global hook. In
there, we have the option to suppress the event, so it doesn't get passed
on to us (or anyt other application). If we don't suppress it, it will
continue on further down the chain, and eventually we will receive the
same event via DirectInput. This means that in practice, we get *two*
events for all shortcut presses. Luckily, the underlying GlobalShortcutEngine
will treat them as one event because they arrive so close to one another.
However, for that mechanism to work, the internal decription of the events
will have to match. We chose to use DirectInput's representation as our
format. Because of that, we convert the incoming keyboard and mouse events
via our Windows hooks into the representation used by DirectInput.