mumble_exe: add config option 'shortcut/windows/uiaccess/enable', allowing users to opt-in to disable UIAccess.
This adds a new config option, 'shortcut/windows/uiaccess/enable' to
Mumble.

This new config option allows users to configure whether or not Mumble
should use UIAccess.

When UIAccess is disabled via the config option, mumble.exe will ensure
it is launched without UIAccess. It does this by re-launching itself
in such a way that its manifest, which specifies uiAccess=true, is
ignored.

No UI for this config option is implemented yet. It is not clear what
should happen in that regard.

Perhaps Mumble should default to not running with UIAccess, and have an
option in the UI to enable it.