os_early: a new mechanism for running OS-specific tasks immediately upon entering main().
This commit introduces a new mechansim for executing OS-specific code
immediately after we enter Mumble's main() procedure.

The entry point for os_early is "os_early_init()" which returns an int.
This return value, if non-zero, will cause Mumble to terminate, using the
return value as the exit status code.

This new system is quite similar to the exisitng os_init() system.
However, it differs in that this system is run very early (upon entering
main()), and it can cause Mumble to terminate, if necessary.

This commit only hooks up os_early on Windows. If other OSes need to
participate in the future, the code in os_win.cpp will need minor
changes.