os_early_win: implement ability to disable UIAccess via config option.
If the setting 'shortcut/windows/uiaccess/enable' is set to false,
Mumble will ensure UIAccess is not enabled for the process.

It does this by re-launching itself in such a way that its manifest,
which specifies uiAccess=true, is ignored. Specifially, we use the
Windows-internal environment variable __COMPAT_LAYER. When this
variable is set to RunAsInvoker, UAC will not be triggered for the
process, and the process will be run as the user that is executing
it. A side-effect of this is that it also disables UIAccess, because
Mumble is unable to elevate itself.