GlobalShortcutWin, MumbleApplication: add suppression support for injected Windows keyboard/mouse message events.
This adds the ability to suppress global shortcuts injected from the main
thread. In practice, this means that Mumble will now honor suppressed keys
*inside* Mumble itself.

This is implemented using a std::promise<bool> in
InjectKeyboardMessageEvent and InjectMouseMessageEvent.

The injectKeyboardMessage and injectMouseMessage methods now
call shouldSuppress() on InjectKeyboardMessageEvent and
InjectMouseMessageEvent to retrieve a std::future<bool> from
the std::promise<bool> of the event. Calling get() on the
std::future<bool> shouldSuppress value blocks the
main thread until the GlobalShortcut engine's thread has set
whether the injected event should be suppressed or not in the
event's std::promise<bool>.

The injectKeyboardMessage and injectMouseMessage methods now return a bool
signalling whether or not the injected event is supposed to be suppressed
or not.

The gswForward method in MumbleApplication is also tweaked to return the
suppression status it receives from injectKeyboardMessage and
injectMouseMessage.

Finally, the event filters in MumbleApplication will now discard keyboard
and mouse messages when the gswForward methods returns true, which signals
that the GlobalShortcut engine told it that the injected event should be
suppressed.