ConnectDialogEdit: layout updates.
Ensure the paste notice label is set to expand vertically, and align
its text at the top.

Promote the qlActions layout to a QWidget, and ensure its margins are
set to 0. Its new name is qwActions.

Apply a horizontal stretch of 4 to the qlPasteNotice label, and a
horizontal stretch of 1 to the qwActions container.

Set the horizontal size policy of both qbDisard and qbFill to be
Expanding, ensuring that the buttons have the same size.

Remove checkmark and cross glyphs from the qbDiscard and qbFill
buttons. They aren't the same size on Windows, which looks odd.
Also, using such glyphs in buttons is not something we usually do,
so I would rather be without it.