GlobalShortcutWin: fall back to 'Unknown' for unknown DirectInput buttons.
When a lookup in qhNames fails, it returns a default-constructed (null and
empty) string.

Do not update the name variable if the result of the lookup is empty.

This ensures we fall back to 'Unknown' when we encounter an unknown
button.

For now, this means that "F13-F24" are shown as K:Unknown instead of just
"K:".