overlay_gl: linux: only assume absolute dynamic entries on glibc.
This fixes a crash in libmumble's library constructor when built against
the musl libc.

The code that locates libc's dlsym(3) relied on addresses in dynamic
entries in the link map returned by dlopen(3) to always be absolute.
This seems to be specific to glibc though, so add a glibc feature test.