OSInfo: ignore deprecation warnings for Gestalt.
Since we now build with a macOS target of 10.8 when
building against Qt 5.10, we get deprecation warnings
when calling gestalt.

The fix, for now, is to ignore deprecation warnings in
OSInfo on macOS.

This should be addressed in a better way in the future.
For that, mumble-voip/mumble#3290 has been filed.