XMLTools: remove default argument 'opstyle' in recurseParse().
This fixes the build on our OS X Universal builder.
Our legacy OS X Universal builder's gcc 4.2 seems to
misparse this.

The only place the default argument is used is in test
code, which is easily changable. Also, this is internal
API, so we can change it as we wish.

Fixes mumble-voip/mumble#3336