SvgIcon: add new class to work around issues with SVG QIcons in Plasma/KStatusNotifierItem
This commit adds new class SvgIcon which allows you to create a QPixmap-backed QIcon
from an SVG.

We need this to fix mumble-voip/mumble#3124.

Basically, SVG-backed QIcons returns an empty list for QIcon::availableSizes().
There is some confusion as to whether this is the expected behavior or not.
(Tracked in mumble-voip/mumble#3374).

However: Plasma/KStatusNotifierItem currently expects QIcon::availableSizes()
to return a valid list of available sizes. For our SVG icons, that is not the case.

This commit works around this by constructing a QIcon backed by pixmaps, such that
QIcon::availableSizes() returns something sensible.

This is implemented via the static method, SvgIcon::addSvgPixmapsToIcon(), which renders
SVG icon in various common sizes (including 22x22 and 44x44, the Plasma notification area
size, and its @2x size).