Only invoke Q_IMPORT_PLUGIN(qico) on Windows.
This fixes a link-time issue for the macOS Universal build.
It only happens there because it's the only Qt 4 static build
we have a build machine for.

The problem is that commit 7a450721f2a76048a57f1ebf9d7d9e89d4bd3db0
changed mumble.pro to only include qico for Windows.

That commit did not update main.cpp to not invoke Q_IMPORT_PLUGIN(qico)
for non-Windows platforms.

This commit fixes that, and hopefully allows the macOS Universal build
to link properly.