Add qtaccessiblewidgets to PLUGINS for Qt 4 builds.
PR mumble-voip/mumble#2688 refactored the .pro file to
avoid special-casing newer versions of Qt.

However, in doing that, it also left out the qtaccessiblewidgets
from PLUGINS for Qt 4 builds.

This commit remedies that, by rewriting the comparison logic to
use integer-based Qt versions instead of complicated logic where
each part of the version (major/minor) is compared separately.

This makes the code easier to read, and fixes a link-time issue
for the macOS Universal build. Since qtaccessiblewidgets is not
on PLUGINS on Qt 4, the build doesn't link against -lqtaccessiblewidgets,
causing linking to fail.