travis-ci: Fix MacOS build for new images
We install the protobuf package via brew, which depends on "python@2".
The build image upgraded the installed "python", and now "python@2" conflicts with it when trying to create symlinks.
We don’t use the symlinked "python" installed by default in the image, so we unlink it to allow the "python@2" package to be installed without conflict.