WASAPI: print log message when the OS blocks access to the microphone
Many users reported that no messages are shown in case Mumble is unable to access the microphone, aside from a simple message with the error code (0x80070005) in the developer console.

This is an issue because since Windows 10 1803 the privacy settings also apply to desktop applications, causing the microphone not to work anymore in many programs (including Mumble) if the user disables access to the microphone globally.