mumble/main.cpp: fix compilation with Qt >= 5.9
When Mumble is statically built, an executable and a shared library are generated.
The executable loads the Mumble library and calls its main entry point.

In order to do that we define the `MumbleMain` function in Mumble, which mimics Qt's WinMain() function.

The Qt qWinMain() function was used to parse command line arguments into a vector. However, this was no longer the case for normal Windows since Qt 5.4, and the function was removed in Qt 5.9.

Since Qt 5.4 the Windows API function CommandLineToArgvW()[1] is used for parsing the command line arguments. Consequently, we now use the same to do the same (this is the effective change of this commit).

The change history and use of qWinMain() in Qt is as follows:

* Before Qt 5.4 the qWinMain() function was used for all Windows platforms.
* Since Qt 5.4 the function was used only for Windows CE (and Windows RT?). Other Windows platforms use the CommandLineToArgvW() Windows API function.
* Qt 5.7 removed support for Windows CE.
* In Qt 5.8 the corresponding Windows CE code path in the main function was removed in https://github.com/qt/qtbase/commit/390598cb4392bbf5f77e8a05578e190b1a9f544a. Consequently qWinMain() was no longer in use but stayed for binary compatibility of the lib module.
* Qt 5.9 removed the now unused qWinMain() function in https://github.com/qt/qtbase/commit/71b54cc24431e8bc6e97f5d62132bd5261764c3a.

[1] https://docs.microsoft.com/en-us/windows/desktop/api/shellapi/nf-shellapi-commandlinetoargvw

---

This change has been tested by building Mumble with MSYS2 and running it on Windows 10.

Qt version: 5.12.1