Add missing advapi32 library reference in server project
win10DisplayableVersion in OSInfo.cpp makes use of the registry to
determine a user displayable windows version string.
This was introduced in 88e664710dd375dc0be42a30dc266efa5833bb4c.

The registry functions used are in the Windows API advapi32 library.
This file was not declared for linkage yet, which leads to a linker
error.