macx/scripts/osxdist.py: copy Opus library into App Bundle
The script copies the CELT (dynamic) libraries into the App Bundle, so that they are installed and then loaded by Mumble.

This commit adds the Opus library to the list of files to copy, necessary because it is loaded at runtime since e0ee016e5c333dabd2603b01decbe8ef2ae030e2.