compiler.pri: disable warnings for deprecated stuff
This commit disables warnings for deprecated stuff, which we are not going to replace due to 1.3.x being frozen.
We also want to keep compatibility with Qt 4, meaning that we would have to adapt each fix from the master branch.

We don't disable the warnings for MSVC because we use our build environment for that.