Implement & auto-select Opus' low delay mode
This commit implements Opus' low delay mode selection, which should result in a big decrease in latency.
The bUseOpusMusicEncoding option has been removed, because now the Opus mode is selected depending on the bitrate. The automatic selection aims to decrease the latency while maintaining acceptable quality.
The exact values may have to be tuned further for a smarter auto-selection because the low delay mode, while causing a strong decrease in latency, requires an higher bitrate to maintain the same quality; in order to compensate, the maximum bitrate in the UI has been increased to 192 kbit/s.

The automatic selection sets:

Low-delay mode when the bitrate is >= 64 kbit/s.
Music mode when the bitrate is >= 32 kbit/s.
VOIP mode when the bitrate is < 32 kbit/s.