scripts: add mumble-version.py
Copied from our current build environment, the header has been changed to match the one in the other scripts.

mumble-version.py determines the full Mumble version based on multiple variables, such as the current commit hash and the latest tag.

There's no reason to keep it in the build environment repository (mumble-releng), because it's supposed to be run inside the Mumble repository and doesn't depend on anything from the build environment.