Remove Speex (audio codec) code
Speex was the first audio codec we used, until it was superseded by CELT in 2009 (Mumble 1.2.0), which in turn was superseded by Opus in 2011 (Mumble 1.2.4).

We retained support for it for 10 years and we believe all active servers are currently using Opus.

This commit removes all references to the Speex API and changes the code so that it ignores a packet if it's encoded with Speex and prints a warning in the console.

To be clear: this commit doesn't remove anything related to the DSP functions, they are part of the SpeexDSP library: it was split from Speex when declared obsolete in favor of Opus.