PulseAudio: Don't send empty data into the stream if there's no audio to send
https://maemo.org/api_refs/5.0/5.0-final/pulseaudio/paplay_8c-example.html

In the example the stream writer callback does nothing and returns in case there is no data to send.

Our code, instead, creates a buffer, fills it with zeroes and writes it to the stream if audio output is not enabled.

This commit changes the callback function's code so that it returns immediately in such case.

The change has been tested by setting an underflow callback function (using pa_stream_set_underflow_callback()) and checking whether underflows occurred.