Remove references to DirectSound in UI
The suggestion to use DirectSound was valid only for Windows versions prior to Vista, because WASAPI was introduced with it.

Fortunately DirectSound didn't appear in the UI as an option in case WASAPI was detected, hopefully we didn't mislead users into thinking that DirectSound is better than WASAPI.

Also, the description was suggesting to use DirectSound on non-Windows operating systems.