Travis CI: transform Linux x86_64 build into ARM64
Travis CI announced support for ARM64 builds a few days ago: https://blog.travis-ci.com/2019-10-07-multi-cpu-architecture-support

We already have a Linux x86_64 build on Azure Pipelines, which also generates an AppImage.

This commit changes the architecture for the build we have on Travis CI, so that it's not redundant anymore (and becomes useful).

The "no-pch" build is not changed because the only differences consist in the PCH header not being used and the SonarQube Scanner being executed.

Also, we don't want to use an ARM64 builder unnecessarily.