Disconnect if server does not send clients user before sync
The protocol requires that a server sends all client information
before sending the sync message that indicates state is matching.
Before this patch violating this condition could bring the client
into a bad state where a connection was established but the client's
user object not existing. Now the client disconnects itself if
such an issue occurs.