JackAudio: remove locks from realtime callbacks, various improvements
According to the JACK specification the process callback cannot have anything that could block.
JACK can also kill the process thread while it is holding the lock which can cause a deadlock.
This prevents both deadlocks in Mumble and xruns in the JACK processes.

There have also been some changes to the order in which the JACK clients get activated.
This is to make sure that they are fully set up before the activate function is called.
This prevents graph errors from showing up in the JACK logs.