Break and fix work queue / various cleanup
The naive type erasing that was being used before was causing it to
never throw an exception when a promise was broken. A new wrapper
class type hides for me and I was able to make the exceptions be thrown.

This means that all the code that looked at the futures needed to have
try/catch around them in case that exception was thrown.

Other changes include the locking in the messaging signaling and
interface. Locks are being held for much shorter periods of time. The
macro for checking user validity has also been reworked as a lambda so
it can be used in both the MurmurGRPCImpl.cpp and Messages.cpp.

Also, I have started work on using boost::optional so that you do not
have to initialize a full protobuf object that you intend to read in
from somewhere else.